fix: type issues

+++ apps/service-scrape/lib-db_service_scrape/migrate.ts
@@ -2,9 +2,13 @@ import { DB_SERVICE_SCRAPE } from './src/index.ts'
 import { LOG } from './src/log.ts'
 import { kyselyPostgresMigrate } from './src/migrator.ts'
 
+if (!DB_SERVICE_SCRAPE) {
+    LOG.fatal('DB_SERVICE_SCRAPE is not defined')
+    process.exit(1)
+}
 const db = await kyselyPostgresMigrate(DB_SERVICE_SCRAPE)
 if (db === null) {
     LOG.fatal('MIGRATION FAILED')
-} else {
-    await db.destroy()
+    process.exit(1)
 }
+await db.destroy()
